+++ app/src/main/java/edu/kmaooad/capstone23/experts/service/ExpertInvitationMailService.java
@@ -1,18 +1,5 @@
 package edu.kmaooad.capstone23.experts.service;
 
-import io.quarkus.mailer.Mail;
-import io.quarkus.mailer.Mailer;
-import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.inject.Inject;
-
-@ApplicationScoped
-public class ExpertInvitationMailService {
-    private static final String INVITATION_SUBJECT = "Capstone23 invitation";
-    @Inject
-    Mailer mailer;
-
-    public void sendInvitationLink(String email, String invitationBody) {
-        mailer.send(Mail.withText(email, INVITATION_SUBJECT, invitationBody));
-    }
-
+public interface ExpertInvitationMailService {
+    void sendInvitationLink(String email, String invitationBody);
 }
